+++ packages/dht/src/dht/routing/RoutingSession.ts
@@ -11,7 +11,7 @@ import { RoutingRpcCommunicator } from '../../transport/RoutingRpcCommunicator'
 import { RecursiveOperationRpcClient, RouterRpcClient } from '../../proto/packages/dht/protos/DhtRpc.client'
 import { toProtoRpcClient } from '@streamr/proto-rpc'
 import { Contact } from '../contact/Contact'
-import { RecursiveOperationRpcRemote } from './RecursiveOperationRpcRemote'
+import { RecursiveOperationRpcRemote } from '../find/RecursiveOperationRpcRemote'
 import { EXISTING_CONNECTION_TIMEOUT } from '../contact/RpcRemote'
 import { getPreviousPeer } from './getPreviousPeer'
